+++ NEWS.md
@@ -22,6 +22,7 @@
 - Fixed: MS files were not always correctly found
 - Fixed: `newProject()` ignored group/blank input for sets workflows
 - Fixed: better error handling for suspect lists with only one valid column
+- Improve suspect list handling when input `mz`/`rt` columns are not numeric
 
 
 # patRoon 2.0